requirements.txt
chromadb
google-generativeai
PyYAML
rich
python-dotenv
